Move IObservableVector implementations to Foundation
