+++ src/Uno.Foundation/Collections/ObservableVectorEnumerableWrapper.cs
@@ -5,7 +5,7 @@ using System.Text;
 using Uno.Extensions.Specialized;
 using Windows.Foundation.Collections;
 
-namespace Windows.UI.Xaml
+namespace Windows.Foundation.Collections
 {
 #if __IOS__
 	[global::Foundation.Preserve(AllMembers = true)]
