+++ src/Uno.Foundation/Collections/ObservableVectorWrapper.cs
@@ -7,7 +7,7 @@ using System.Text;
 using Uno.Extensions;
 using Windows.Foundation.Collections;
 
-namespace Windows.UI.Xaml
+namespace Windows.Foundation.Collections
 {
 #if __IOS__
 	[global::Foundation.Preserve(AllMembers = true)]
